refactor(cache): isolate KV cache event publishers
