+++ rtp_llm/cpp/cache/BUILD
@@ -77,21 +77,10 @@ cc_library(
     ] + torch_deps(),
 )
 
-cc_library(
+alias(
     name = "kv_cache_event_publisher",
-    srcs = [
-        "KVCacheEventPublisher.cc",
-    ],
-    hdrs = [
-        "KVCacheEventPublisher.h",
-    ],
-    copts = copts(),
+    actual = "//rtp_llm/cpp/cache/events:kv_cache_event_publisher",
     visibility = ["//visibility:public"],
-    deps = [
-        "//rtp_llm/cpp/utils:core_utils",
-        "@curl//:curl",
-        "@rapidjson//:rapidjson",
-    ],
 )
 
 cc_library(
@@ -114,7 +103,7 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         ":cache_types",
-        ":kv_cache_event_publisher",
+        "//rtp_llm/cpp/cache/events:kv_cache_event",
         "//rtp_llm/models_py/bindings/core:exec_ops_hdr",
         "//rtp_llm/cpp/disaggregate/cache_store",
         "//rtp_llm/cpp/engine_base/stream:complete_token_ids",
@@ -268,7 +257,7 @@ cc_library(
     visibility = ["//visibility:public"],
     deps = [
         ":cache_core",
-        ":kv_cache_event_publisher",
+        "//rtp_llm/cpp/cache/events:kv_cache_event_publisher",
         "//rtp_llm/cpp/cache/connector",
         "//rtp_llm/models_py/bindings/core:exec_ops_hdr",
         "//rtp_llm/cpp/utils:profiling_scope",
