+++ rtp_llm/cpp/cache/test/BUILD
@@ -91,19 +91,6 @@ cc_test(
 )
 
 cc_test(
-    name = "kv_cache_event_publisher_test",
-    srcs = [
-        "KVCacheEventPublisherTest.cc",
-    ],
-    copts = test_copts,
-    deps = [
-        "//rtp_llm/cpp/cache:kv_cache_event_publisher",
-        "@com_google_googletest//:gtest",
-        "@com_google_googletest//:gtest_main",
-    ],
-)
-
-cc_test(
     name = "block_pool_test",
     srcs = [
         "BlockPoolTest.cc",
